Put BeanDeserializer to Deserializer
