+++ common/src/main/java/org/astraea/common/consumer/Deserializer.java
@@ -16,8 +16,12 @@
  */
 package org.astraea.common.consumer;
 
+import java.util.Arrays;
 import java.util.Base64;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 import org.apache.kafka.common.header.Headers;
 import org.apache.kafka.common.serialization.ByteArrayDeserializer;
 import org.apache.kafka.common.serialization.DoubleDeserializer;
@@ -28,6 +32,7 @@ import org.apache.kafka.common.serialization.StringDeserializer;
 import org.astraea.common.Header;
 import org.astraea.common.json.JsonConverter;
 import org.astraea.common.json.TypeRef;
+import org.astraea.common.metrics.BeanObject;
 
 @FunctionalInterface
 public interface Deserializer<T> {
@@ -72,6 +77,7 @@ public interface Deserializer<T> {
   Deserializer<Long> LONG = of(new LongDeserializer());
   Deserializer<Float> FLOAT = of(new FloatDeserializer());
   Deserializer<Double> DOUBLE = of(new DoubleDeserializer());
+  Deserializer<BeanObject> BEAN_OBJECT = new BeanDeserializer();
 
   /**
    * create Custom JsonDeserializer
@@ -98,6 +104,35 @@ public interface Deserializer<T> {
       else {
         return jackson.fromJson(Deserializer.STRING.deserialize(topic, headers, data), typeRef);
       }
+    }
+  }
+
+  /**
+   * Deserialize byte arrays to string and then parse the string to `BeanObject`. It is inverse of
+   * BeanObject.toString().getBytes().
+   */
+  class BeanDeserializer implements Deserializer<BeanObject> {
+    @Override
+    public BeanObject deserialize(String topic, List<Header> headers, byte[] data) {
+      var beanString = new String(data);
+      Pattern p =
+          Pattern.compile("\\[(?<domain>[^:]*):(?<properties>[^]]*)]\n\\{(?<attributes>[^}]*)}");
+      Matcher m = p.matcher(beanString);
+      if (!m.matches()) return null;
+      var domain = m.group("domain");
+      var propertiesPairs = m.group("properties").split("[, ]");
+      var attributesPairs = m.group("attributes").split("[, ]");
+      var properties =
+          Arrays.stream(propertiesPairs)
+              .map(kv -> kv.split("="))
+              .filter(kv -> kv.length >= 2)
+              .collect(Collectors.toUnmodifiableMap(kv -> kv[0], kv -> kv[1]));
+      var attributes =
+          Arrays.stream(attributesPairs)
+              .map(kv -> kv.split("="))
+              .filter(kv -> kv.length >= 2)
+              .collect(Collectors.toUnmodifiableMap(kv -> kv[0], kv -> (Object) kv[1]));
+      return new BeanObject(domain, properties, attributes);
     }
   }
 }
